3/26 sat I worked
Chore : atoi local variable modified.

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: wochae <wochae@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/19 21:16:33 by wochae            #+#    #+#             */
-/*   Updated: 2022/03/25 19:13:51 by wochae           ###   ########.fr       */
+/*   Updated: 2022/03/26 16:45:41 by wochae           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	ft_atoi(const char *str)
 {
-	long long		minus;
-	long long		ret;
+	int			minus;
+	long		ret;
 
 	ret = 0;
 	minus = 1;
@@ -31,7 +31,8 @@ int	ft_atoi(const char *str)
 			&& ((ret == 922337203685477580 && *str > '7')
 				|| ret > 922337203685477580) && ret < 1000000000000000000)
 			return (-1);
-		if (minus == -1 && ((ret == 922337203685477580 && *str > '8')
+		if (minus == -1
+			&& ((ret == 922337203685477580 && *str > '8')
 				|| ret > 922337203685477580) && ret < 1000000000000000000)
 			return (0);
 		ret *= 10;
